[sync]

+++ feature/countdown/src/main/java/ir/kasebvatan/countdown/model/CountdownState.kt
@@ -4,5 +4,5 @@ package ir.kasebvatan.countdown.model
 data class CountdownState(
     val counterState: CounterState = CounterState.INITIAL,
     val workingState: WorkingState = WorkingState.REST,
-    val remainTime: Long = WORKING_DURATION,
+    val remainTime: Int = WORKING_DURATION,
 )
