+++ library/ext-functions/src/main/java/ir/kasebvatan/ext_functions/Time.kt
@@ -1,10 +1,10 @@
 package ir.kasebvatan.ext_functions
 
-val Long.minutes: Long
+val Int.minutes: Int
     get() = this / 60
 
 
-val Long.seconds: Long
+val Int.seconds: Int
     get() = this % 60
